refactor(framework): Deprecate NDArray/Parameters conversion functions
