+++ framework/py/flwr/compat/common/parameter.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Parameter conversion."""
+"""Legacy parameter conversion."""
 
 
 from io import BytesIO
