+++ framework/py/flwr/compat/common/parameter_test.py
@@ -12,10 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""...
-
-tests.
-"""
+"""Legacy parameter conversion tests."""
 
 
 import numpy as np
